% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderule-package.R
\docType{package}
\name{ponderule}
\alias{ponderule}
\alias{ponderule-package}
\title{ponderule:  Full Text Search of 'Kindai-tanka'}
\description{
Full text search functionality for a 'Kindai-tanka' corpus.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/ponderule}
  \item Report bugs at \url{https://github.com/paithiov909/ponderule/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Authors:
\itemize{
  \item Yuna Murata (Maintainer of the corpus)
}

}
\keyword{internal}
