% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderule-package.R
\docType{data}
\name{tanka}
\alias{tanka}
\title{Corpus of kindai tanka}
\format{
An object of class \code{data.frame} with 25833 rows and 6 columns.
}
\usage{
tanka
}
\description{
Corpus of kindai tanka
}
\keyword{datasets}
