% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderule-package.R
\docType{data}
\name{wakati}
\alias{wakati}
\title{Pre-tokenized text from 'tanka'}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25832 rows and 2 columns.
}
\usage{
wakati
}
\description{
Pre-tokenized text from 'tanka'
}
\keyword{datasets}
